Refactor: Packaging(course)
